+++ hooks/useToast/useToast.ts
@@ -1,52 +1,53 @@
-import React, { useRef, useState, useEffect } from 'react';
-import * as ReactDOMClient from 'react-dom/client';
-import { nanoid } from 'nanoid';
-import { Toast } from 'components';
-import { IToast } from 'types';
+export {};
+// import React, { useRef, useState, useEffect } from 'react';
+// import * as ReactDOMClient from 'react-dom/client';
+// import { nanoid } from 'nanoid';
+// import { Toast } from 'components';
+// import { IToast } from 'types';
 
-interface IUseToastProps {
-  top: number;
-  right: number;
-}
+// interface IUseToastProps {
+//   top: number;
+//   right: number;
+// }
 
-export function useToast({ top, right }: IUseToastProps) {
-  const portalId = 'toast-portal';
+// export function useToast({ top, right }: IUseToastProps) {
+//   const portalId = 'toast-portal';
 
-  const portalElement = useRef<HTMLDivElement | null>(null);
-  const [toasts, setToasts] = useState<Array<IToast>>([]);
-  const [root, setRoot] = useState<ReactDOMClient.Root | null>(null);
+//   const portalElement = useRef<HTMLDivElement | null>(null);
+//   const [toasts, setToasts] = useState<Array<IToast>>([]);
+//   const [root, setRoot] = useState<ReactDOMClient.Root | null>(null);
 
-  const createToast = ({ status, message, duration }: Omit<IToast, 'key'>) => {
-    setToasts((prev) => [
-      ...prev,
-      { key: nanoid(), status, message, duration },
-    ]);
-  };
+//   const createToast = ({ status, message, duration }: Omit<IToast, 'key'>) => {
+//     setToasts((prev) => [
+//       ...prev,
+//       { key: nanoid(), status, message, duration },
+//     ]);
+//   };
 
-  const removeToast = ({ key }: { key: string }) => {
-    setToasts((prev) => prev.filter((toast) => toast.key !== key));
-  };
+//   const removeToast = ({ key }: { key: string }) => {
+//     setToasts((prev) => prev.filter((toast) => toast.key !== key));
+//   };
 
-  useEffect(() => {
-    if (portalElement.current) {
-      root!.render(
-        React.createElement(Toast, {
-          toasts,
-          removeToast,
-          top,
-          right,
-        }),
-      );
+//   useEffect(() => {
+//     if (portalElement.current) {
+//       root!.render(
+//         React.createElement(Toast, {
+//           toasts,
+//           removeToast,
+//           top,
+//           right,
+//         }),
+//       );
 
-      return;
-    }
+//       return;
+//     }
 
-    const toastDOM = document.getElementById(portalId);
+//     const toastDOM = document.getElementById(portalId);
 
-    portalElement.current = toastDOM as HTMLDivElement;
+//     portalElement.current = toastDOM as HTMLDivElement;
 
-    setRoot(ReactDOMClient.createRoot(portalElement.current));
-  }, [toasts]);
+//     setRoot(ReactDOMClient.createRoot(portalElement.current));
+//   }, [toasts]);
 
-  return { createToast };
-}
+//   return { createToast };
+// }
